Fix precision errors in visibility pipeline
When storing floats in textures, we store them as 16-bit floats.
This can cause a noticeable disparity if we render terrain using
floats that didn't go through a texture (floats that were always 32-bit)
vs. if we use floats read from a texture.

This was most visible in the visibility pipeline, though it
affected the deferred one as well. It also mainly affected
world-space coordinates, since colors and UVs are small enough in
value (components are in [0, 1]) to be only minimally affected by this.

The fix for this is to temporarily convert the floats to a smaller value
when storing them in a texture, then converting them back to their true
value after reading them from the texture. This still incurs a loss of
precision, but not as much as when the floats grow unboundedly.