Fix flickering issue when precision fix is active
This was caused by asynchronous camera updates. Since the precision fix
relies on the camera position to re-compute the fragments' real
world-space position, if the camera position is not consistently
updated for all stages of the pipelines, weird things would happen where
the world position of all fragments would be shifted by tileDim
momentarily, causing normal computations to go horribly wrong.

Camera position is now updated through a VkCmdCopyBuffer() call, which
effectively allows us to synchronize this update. Previously, it was
done via memcpy to mapped memory, which meant that the data was only
transferred to the GPU at some indeterminate time. This is fine, for
example, for the time buffer, since small variations in the time will be
imperceptible, but the way the camera buffer was being used led to
flickering when it transitioned between multiples of tileDim.